Add static backdrop, most modalSave to locationSave
