+++ app/components/location_modal_component.html.haml
@@ -1,4 +1,4 @@
-.p-0.modal.fade{id: "#{@component}sModal", "aria-hidden" => "true", "aria-labelledby" => "#{@component}sModal", :role => "dialog", :tabindex => "-1" }
+.p-0.modal.fade{id: "#{@component}sModal", "data-backdrop": "static", "aria-hidden" => "true", "aria-labelledby" => "#{@component}sModal", :role => "dialog", :tabindex => "-1" }
   .modal-dialog.modal-dialog-centered{:role => "document"}
     .modal-content
       .modal-header
